exer 8 implemented. Added proper Doc comments
