+++ martin/02-oop-inheritance/js/Actor.js
@@ -1,0 +1,13 @@
+'use strict';
+/*
+ * Exercise 7
+ * @class Actor
+ * @constructor
+ */
+class Actor {
+  constructor(name, age) {
+    this.name = name;
+    this.age = age;
+  }
+}
+export { Actor }
