+++ martin/02-oop-inheritance/js/Logger.js
@@ -1,0 +1,17 @@
+'use strict';
+/*
+ *Exercise 5
+ * @class Logger */
+class Logger {
+  constructor() {}
+    /*
+     * @param {string} info 
+     * 
+     * @memberOf Logger
+     */
+  log(info) {
+    console.log(`The ${info} has been emmited`);
+  }
+}
+
+export { Logger }
